% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node_attr.R
\name{add_node_attr}
\alias{add_node_attr}
\title{Add node attributes}
\usage{
add_node_attr(node_layout, node_data)
}
\arguments{
\item{node_layout}{Data frame of node layout returned by \code{get_layout()}.}

\item{node_data}{Data frame or tibble of node information. Must have at least one column
called "name" for node names to join by. To change labels that appear on nodes, include
a column with the name "label".}
}
\value{
A tibble.
}
\description{
Adds additional information to the node e.g. variable to colour by
}
