% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteris_exclusio_ggconsort.R
\name{criteris_exclusio_ggconsort}
\alias{criteris_exclusio_ggconsort}
\title{Aplica criteris partir d'una llista de n criteris d'exclusió que estan en un fitxer extern (Tipo excel)}
\usage{
criteris_exclusio_ggconsort(
  dt = dades,
  taulavariables = "VARIABLES_R3b.xls",
  criteris = "exclusio1",
  missings = T,
  ...
)
}
\arguments{
\item{dt}{dataframe/tibble}

\item{taulavariables}{string referent a path o tibble on hi ha una columna consten les exclusions: "conductor_cars.xls"}

\item{criteris}{string referent a la columna on consten les exclusions}
}
\description{
Aplica criteris partir d'una llista de n criteris d'exclusió que estan en un fitxer extern (Tipo excel)
}
