% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_dummies_exclusions.R
\name{generar_dummies_exclusions}
\alias{generar_dummies_exclusions}
\title{Genera n variables dummis (exclusio1,exclusio2,....exclusion) segons criteris d'exclusio determinats}
\usage{
generar_dummies_exclusions(
  dt = iris,
  criteris = c("Sepal.Length < 5", "Species=='setosa'")
)
}
\arguments{
\item{dt}{dataframe/tibble}

\item{criteris}{vector string amb la definició del criteris: criteris=c("Sepal.Length < 5","Species=='setosa'")}
}
\value{
data.frame o tibble amb n columnes tipo dummie (0/1) afegides referides a les n exclusions
}
\description{
Genera n variables dummis (exclusio1,exclusio2,....exclusion) segons criteris d'exclusio determinats
}
\examples{

generar_dummies_exclusions(dt=iris,criteris=c("Sepal.Length < 5","Species=='setosa'"))



}
