% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edges.R
\name{get_edges}
\alias{get_edges}
\title{Get edges}
\usage{
get_edges(data, plot_nodes)
}
\arguments{
\item{data}{Data frame or tibble of edges. Must have at least two columns, if not labelled "from"
and "to" it is assumed first column are "from" node names,
second column is "to" node names. Node names must be unique.}

\item{plot_nodes}{Tibble of node coordinates output from \code{get_nodes()}.}
}
\value{
A tibble.
}
\description{
Define the end points of the arrows
}
