% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layout.R
\name{get_layout}
\alias{get_layout}
\title{Get layout}
\usage{
get_layout(data)
}
\arguments{
\item{data}{Data frame or tibble of edges. Must have at least two columns, if not labelled "from"
and "to" it is assumed first column are "from" node names,
second column is "to" node names. Node names must be unique.}
}
\value{
A tibble.
}
\description{
Obtains the coordinates for the nodes of the flowchart from the input edges
}
