% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{Get nodes}
\usage{
get_nodes(node_layout, x_nudge = 0.35, y_nudge = 0.25)
}
\arguments{
\item{node_layout}{Data frame of node layout returned by \code{get_layout()}}

\item{x_nudge}{Distance from centre of edge of node box in x direction. Default 0.35.}

\item{y_nudge}{Distance from centre of edge of node box in y direction. Default 0.25.}
}
\value{
A tibble.
}
\description{
Define the corners of the nodes.
}
