% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggflowchart.R
\name{ggflowchart}
\alias{ggflowchart}
\title{Generate a flowchart in ggplot2}
\usage{
ggflowchart(data, node_data = NULL)
}
\arguments{
\item{data}{Data frame or tibble of edges. Must have two columns, first column are "from" node names,
second column is "to" node names. Node names must be unique.}

\item{node_data}{Data frame or tibble of node information. Must have at least one column
called "name" for node names to join by. Default NULL.}
}
\value{
A ggplot2 object.
}
\description{
Generates the flowchart
}
