% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_conductor.R
\name{read_conductor}
\alias{read_conductor}
\title{Read_conductor}
\usage{
read_conductor(fitxer, ...)
}
\arguments{
\item{fitxer}{Fitxer que el convertirem a tibble.}

\item{...}{Altres funcions}
}
\value{
Una taula tibble
}
\description{
Llegeix el fitxer conductor de diferents formats txt o rds o xls xlsx o data_frame tibble
}
