% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggflowchart.R
\docType{package}
\name{ggflowchart}
\alias{ggflowchart}
\title{ggflowchart Package}
\description{
ggflowchart contains functions to create flow-chart from ggconsort functions.
It is not require to know ggconsort functions because ggflowchart create automatically
}
\author{
Jordi Real \email{-jordireal@gmail.com-}
}
