% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_conductor.R
\name{read_conductor}
\alias{read_conductor}
\title{Read conductor file different formats txt o rds o xls xlsx o data_frame tibble}
\usage{
read_conductor(fitxer)
}
\arguments{
\item{fitxer}{Character as a file name and path or data.frame}
}
\description{
Read conductor file different formats txt o rds o xls xlsx o data_frame tibble
}
